Fix FreeRTOS features
Use of FreeRTOS heap management is possible

Fix stack in the FreeRTOS handler

gic handler will no longer end the interrupt of EPIT and FreeRTOS handler will do

Cleaning files

+++ core/kernel/FreeRTOS/heap_4.c
@@ -59,7 +59,7 @@
 
 /* The application writer has already defined the array used for the RTOS
 * heap - probably so it can be placed in a special segment or address. */
-    extern uint8_t ucHeap[ configTOTAL_HEAP_SIZE ];
+    extern uint32_t ucHeap[ configTOTAL_HEAP_SIZE ];
 #else
     PRIVILEGED_DATA static uint32_t ucHeap[ configTOTAL_HEAP_SIZE ];
 #endif /* configAPPLICATION_ALLOCATED_HEAP */
